♻️ refactor(mcp-local-analyze-template-search): move apply_style to Engine, fix grid/apply_style duplication
- Add apply_style() public function to sdpm.api (Engine is source of truth)
- Add thin wrapper in tools.py, server.py calls it via tools
- Remove duplicate apply_style from server_acp.py (mcp instance is shared)
- Remove duplicate grid from server_acp.py, unify signature in server.py
- grid signature now matches ACP version: purpose (required) before spec

SPEC: fix/mcp-local-analyze-template-search
